requirements.txt
requests~=2.25.1
Pillow~=7.0.0
beautifulsoup4~=4.10.0
python-escpos~=2.2.0
luma.oled==3.8.1
luma.core==2.3.1
html5lib~=1.1
python-escpos~=2.2.0